split libjinue, simplify includes
