+++ include/jinue/util.h
@@ -1,22 +1,22 @@
 /*
- * Copyright (C) 2019 Philippe Aubertin.
+ * Copyright (C) 2022 Philippe Aubertin.
  * All rights reserved.
 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- * 
+ *
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 
+ *
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,55 +29,21 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <jinue/errno.h>
-#include <jinue/ipc.h>
-#include <jinue/syscall.h>
+#ifndef _JINUE_UTIL_H
+#define _JINUE_UTIL_H
 
-intptr_t jinue_send(
-        int                      fd,
-        intptr_t                 function,
-        const jinue_message_t   *message,
-        int                     *perrno) {
+#include <stdarg.h>
 
-    jinue_syscall_args_t args;
+void jinue_vinfo(const char *restrict format, va_list arg);
 
-    args.arg0 = (uintptr_t)function;
-    args.arg1 = (uintptr_t)fd;
-    args.arg2 = (uintptr_t)message;
-    args.arg3 = 0;
+void jinue_vwarning(const char *restrict format, va_list arg);
 
-    return jinue_syscall_with_usual_convention(&args, perrno);
-}
+void jinue_verror(const char *restrict format, va_list arg);
 
-intptr_t jinue_receive(int fd, const jinue_message_t *message, int *perrno){
-    jinue_syscall_args_t args;
+void jinue_info(const char *restrict format, ...);
 
-    args.arg0 = JINUE_SYS_RECEIVE;
-    args.arg1 = (uintptr_t)fd;
-    args.arg2 = (uintptr_t)message;
-    args.arg3 = 0;
+void jinue_warning(const char *restrict format, ...);
 
-    return jinue_syscall_with_usual_convention(&args, perrno);
-}
+void jinue_error(const char *restrict format, ...);
 
-intptr_t jinue_reply(const jinue_message_t *message, int *perrno) {
-    jinue_syscall_args_t args;
-
-    args.arg0 = JINUE_SYS_REPLY;
-    args.arg1 = 0;
-    args.arg2 = (uintptr_t)message;
-    args.arg3 = 0;
-
-    return jinue_syscall_with_usual_convention(&args, perrno);
-}
-
-int jinue_create_ipc(int flags, int *perrno) {
-    jinue_syscall_args_t args;
-
-    args.arg0 = JINUE_SYS_CREATE_IPC;
-    args.arg1 = (uintptr_t)flags;
-    args.arg2 = 0;
-    args.arg3 = 0;
-
-    return jinue_syscall_with_usual_convention(&args, perrno);
-}
+#endif
