+++ include/kernel/types.h
@@ -33,10 +33,10 @@
 #define JINUE_KERNEL_TYPES_H
 
 #include <jinue/shared/ipc.h>
-#include <jinue/shared/list.h>
 #include <jinue/shared/syscall.h>
 #include <jinue/shared/types.h>
 #include <kernel/i686/types.h>
+#include <kernel/list.h>
 #include <stddef.h>
 #include <stdint.h>
 #include <stdbool.h>
